style: Move package and delete package
