--- src/test/java/coordinate/Figure.java
@@ -1,13 +1,0 @@
-package coordinate;
-
-import java.util.List;
-
-public interface Figure {
-    List<Point> getPoints();
-
-    int size();
-
-    String getName();
-
-    double area();
-}
