--- src/test/java/coordinate/FigureCreator.java
@@ -1,7 +1,0 @@
-package coordinate;
-
-import java.util.List;
-
-public interface FigureCreator {
-    Figure create(List<Point> points);
-}
