+++ src/test/java/webserver/http/headerfields/HttpMethodTest.java
@@ -1,8 +1,7 @@
-package webserver.http;
+package webserver.http.headerfields;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.http.headerfields.HttpMethod;
 
 import static org.assertj.core.api.Assertions.assertThat;
