+++ src/test/java/webserver/http/headerfields/HttpVersionTest.java
@@ -1,8 +1,7 @@
-package webserver.http;
+package webserver.http.headerfields;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.http.headerfields.HttpVersion;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
